+++ cw-vault-standard/CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `base_token` and `vault_token` fields to `VaultContract` helper struct.
 - Remove faulty usage of `to_binary` in `VaultContract` query helper functions.
 - Replace usage of deprecated `to_binary` with `to_json_binary`.
+- Deprecated `VaultStandardQueryMsg::PreviewDeposit` and `VaultStandardQueryMsg::PreviewRedeem`.
+    - There queries turned out to be too difficult to implement in most cases. We recommend to use transaction simulation from non-contract clients such as frontends.
 
 ## [0.3.3] - 2023-09-27